Adds ability to also set "on" state of lamps
With this we can now turn of lamps at certain times of day, or ensure that a turned of light (via app) is turned on again.
If the "on" property is not defined, the app does not change a lights on state, but waits with until the user turns it on again, and then sets the defined light state.