Updates error handling of API calls, retrying on failures
Now we throw:
- BridgeConnectionFailure if the HTTP call failed
- BridgeAuthenticationFailure if the bridge rejected the username
- HueApiFailure if the API response contained and error

We now check for connection and valid username during startup, to abort on authentication issues immediately.
We also check for these exceptions during execution and retry the state until it succeeds.
This behaviour might change in the future based on real world feedback.
Aborting execution should however be the last resort, to not leave users in the dark (literally).