Adds new multi_colorloop effect
This is an advanced version of the colorloop for groups, which offsets the color loop for each light in the group so that they loop in different colors.
This is achieved by turning the lights in the group on and off again, with a defined pause between each light. Resulting in different loop start times, and therefore different colors.