WIP for Home Assistant API integration
Refactor off tracking to support HASS api: Now we check the off state before we make any light updates. This is needed as the HASS api only provides a turn_on command, which automatically adds the "on:true" property. But we don't always want to turn on the lights, unless explicitly set.