Add Scene Sync for Home Assistant (#23)
* Improve HassEventHandler to better ignore unwanted events

* Add scene sync for Home Assistant

Introduced HassAreaRegistry for managing Home Assistant area-light relationships and caching. Adapted SceneEventListenerImpl to use dynamic scene matching.

* Add --require-scene-activation to HueScheduler

This flag prevents triggering Hue Scheduler when lights are turned on, except for synced scenes. 

* Refactor reschedule and power-on logic to improve scene sync reliability

* Don't ignore states if they have force:true

* Update Hue event handler to also consider the same scene activated again

* Don't override synced-flag when lights are turned on physically

* Ignore scenes with no name in SyncedScene check

* Improve HA availability check during startup

Now we wait for the "homeassistant_started" event if HA is not fully available yet.
Still room for improvement.

* Improve change detection for HA events

Events are not detected as turned on anymore when just their properties change.
Lights are not detected as turned on anymore if their previous state is null.

* Improve connection handling for HassEventStreamReader

Now we also listen for "homeassistant_started" events.

* Make ResourceNotFoundException of type ApiFailure

This way, requests are correctly retried when HA is currently restarting.
Since there might be a time frame when HA is answering requests, but does not yet have all entities loaded.

* Add cache clear on HA start/restart

* Use light off event to reset turnedOnBySyncedScene flag

Instead of resetting it on the turn-on event. Since those events might be detected delayed for bigger groups and may cause false positives when a synced scene is turned on.

* Increase default value for sceneActivationIgnoreWindow

Adjusts default delay from 5 to 8 seconds to reduce false positives caused by delayed events in larger groups after scene activations.

* Automatically re-sync scenes on HA restarts

This is needed as synced scenes are only temporary in HA and are cleared on each restart.

* Reset justTurnedOn after timeout

This fixes issues where synced scene activation sets this flag for lights that are currently off. Resulting in states being incorrectly applied later on, even though they are actually off.
Default timout set to 10 seconds.
Added "--just-turned-on-window" flag and environment variable to configure the default timout.

* Optimize getAffectedIdsByScene for HueApi by ignoring off lights

* Use Set for affectedIdsForScene to avoid repeated turnOns

* Remove "just turned on" logic from ManualOverrideTracker

Removes the logic and related methods for tracking if a light was "just turned on." We now just use the power on flag of the snapshots for detecting lights being turned on. So we don't need potentially flaky timed-based caching.

* Add rate limiting to GET requests for Hue bridges (#20)

Remove the confusing log message for HTTP 429 responses, simplifying the error message.

* Update dependencies

* Improve thread safety of LightEventListenerImpl

* Ignore any potential duplicate area, entity and device entries

* Remove unused code and improve timing consistency

By reusing the captured current time across multiple calculations. This way we avoid any time drifts.
We only use the up-to-date time when calculating delays for scheduling purposes.
Improve transition truncation to also consider seconds.

* Remove obsolete isSkipSceneSync flag

A recent refactoring to how power-on states are created and scheduled made this flag obsolete.

* Add tests for HassAvailabilityListener and assertConnection

* Bump version to 0.13.0

* Update logo

* Increase thresholds for comparisons of bri and ct

* Add some trace logging for overridden lights or groups

* Make manual override thresholds configurable

Set better defaults for brightness, ct, and color.

* Adjust default color override threshold

* Ignore unreachable lights during group tracking

Adjusted logic to skip unreachable lights when comparing group modifications.

* Determine availability directly via API

Add unavailable property to LightState.
Ignore any modifications for unavailable lights.
This prevents false positive manual modification detections.

* Add resource modification tracking for HassApi

To update the cache instead of clearing it. This prevents too many state API fetches.

* Correctly abort startup for HASS on a BridgeAuthenticationFailure

* Add some additional parameter value checks on startup

* Remove other properties for off states in HA synced scenes

HA does not support e.g. setting an individual transition time inside scenes.

* Replace `isEmpty` with `isBlank` in input parsing

This ensures that lines containing only whitespace are filtered out.

* Simplify README and docs

* Restructure CLI flag documentation